+++ src/components/days-list.js
@@ -1,13 +1,6 @@
 import AbstractComponent from '../components/abstract-component';
 
-const createDaysListTemplate = () => {
-  return (
-    `<ul class="trip-days">
-
-      </ul>
-    `
-  );
-};
+const createDaysListTemplate = () => `<ul class="trip-days"></ul>`;
 
 export default class DaysList extends AbstractComponent {
   getTemplate() {
