+++ src/components/points-list.js
@@ -1,13 +1,6 @@
 import AbstractComponent from '../components/abstract-component';
 
-const createDayPointListTemplate = () => {
-  return (
-    `<ul class="trip-events__list">
-
-      </ul>  
-    `
-  );
-};
+const createDayPointListTemplate = () => `<ul class="trip-events__list"></ul>`;
 
 export default class PointsList extends AbstractComponent {
   getTemplate() {
